5 tasks for 1 hour lazy evening

+++ 3.4/C.py
@@ -3,5 +3,6 @@ a, b, c = map(float, input().split())
 for value in count(a, c):
     if value > b:
         break
-    print(float(round(value, 2)))
+    print(round(value, 2))
 
+# 5 mins
